♻️️ Refactor : "SeasonalCocktail 도메인 패키지 및 코드 구조 조정"
- Java Docs 작업 진행

- 컨트롤러 및 서비스 정의 명확성과 가독성을 위한 리팩토링 진행

Ref : #5, #84
