--- src/main/java/alcoholboot/toastit/feature/calendarcocktail/repository/CalendarCocktailRepository.java
@@ -1,8 +1,0 @@
-package alcoholboot.toastit.feature.calendarcocktail.repository;
-
-import alcoholboot.toastit.feature.calendarcocktail.entity.CalendarCocktail;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CalendarCocktailRepository extends JpaRepository<CalendarCocktail, Long> {
-
-}
