+++ src/main/java/alcoholboot/toastit/feature/localcocktail/entity/LocationCocktail.java
@@ -9,7 +9,7 @@ import lombok.Getter;
  */
 @Getter
 @Entity
-@Table(name = "location_cocktail")
+@Table(name = "location_cocktails")
 public class LocationCocktail {
 
     @Id
