+++ src/main/java/alcoholboot/toastit/feature/seasonalcocktail/repository/SeasonalCocktailRepository.java
@@ -1,0 +1,7 @@
+package alcoholboot.toastit.feature.seasonalcocktail.repository;
+
+import alcoholboot.toastit.feature.seasonalcocktail.entity.SeasonalCocktail;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SeasonalCocktailRepository extends JpaRepository<SeasonalCocktail, Long> {
+}
