+++ src/main/resources/templates/feature/seasonalcocktail/recipe-preview.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org" lang="ko">
 <head>
-    <title>Date Page</title>
+    <title>Seasonal Cocktail | ToastIt</title>
     <link th:href="@{/css/fragment/navbar-fragment.css}" rel="stylesheet">
     <style>
         .container {
