+++ src/main/resources/templates/index.html
@@ -284,7 +284,7 @@
                     </a>
                 </div>
                 <div class="swiper-slide">
-                    <a th:href="@{/season-date}">
+                    <a th:href="@{/season}">
                         <img src="/image/dateCocktail.png" alt="Image">
                         <div class="text">날짜</div>
                     </a>
